feat(move-give-feedback): move give feedback to more prominent
